feat: add buffered point and polyline options to shape filter
Closes #697
